Modified withAuthenticate
I needed to refer to the solution code for this one. I added an actual state object (as I thought I would need to eventually) called loggedin and then I used a componentDidMount to check and edit the state of the loggedin state object. On top of this, a simple if funciton that cehcks for the boolen value of the loggIn state value and depending on that value it would return a specific component.